% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activateVirtualKeyboard.R
\name{activateVirtualKeyboard}
\alias{activateVirtualKeyboard}
\title{Function that activates a virtual keyboard}
\usage{
activateVirtualKeyboard(keyboard)
}
\arguments{
\item{keyboard}{list of js/css scripts generated with generateVirtualKeyboard() function}
}
\description{
Function that activates a virtual keyboard
}
\examples{
activateVirtualKeyboard("text")

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
