% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVirtualKeyboard.R
\name{generateVirtualKeyboard}
\alias{generateVirtualKeyboard}
\title{Function that generates and loads the necessary js/css scripts for the virtual keyboard to work}
\usage{
generateVirtualKeyboard(textIds)
}
\arguments{
\item{textIds}{character or list with characters, indicating the id's of the textInputs where the virtual keyboard should be applied}
}
\description{
Function that generates and loads the necessary js/css scripts for the virtual keyboard to work
}
\examples{
keyboard <- generateVirtualKeyboard(textIds)

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
