% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboardInput.R
\name{keyboardInput}
\alias{keyboardInput}
\title{An SVG Keyboard

This input operates like a \code{shiny::radioButtons()} where you can select keys to depress to simulate typing!}
\usage{
keyboardInput(inputId = ".id", color_palette = "sharla1", selected = NULL)
}
\arguments{
\item{color_palette}{the color palette to use options include:}

\item{selected}{the keys selected by default}

\item{id}{Shiny input id}
}
\value{

}
\description{
An SVG Keyboard

This input operates like a \code{shiny::radioButtons()} where you can select keys to depress to simulate typing!
}
\examples{
\dontrun{
keyboardInput("keebs", color_palette = "sharla1")
}
}
