% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateKeyboard.R
\name{updateKeyboard}
\alias{updateKeyboard}
\title{Function that generates js to bind keyboard to textIds that are generated in the serverside}
\usage{
updateKeyboard(textIds)
}
\arguments{
\item{textIds}{character or list with characters, indicating the id's of the textInputs where the virtual keyboard should be applied}
}
\description{
Function that generates js to bind keyboard to textIds that are generated in the serverside
}
\examples{
observeEvent(input$text, {
    shinyjs::runjs(updateKeyboard("text"))
 })

}
\author{
Emelie Hofland, \email{emelie_hofland@hotmail.com}
}
